Fix oer-reveal-add-alternate-types
Only add footnote in title for LaTeX export.
For HTML export, the footnote code would be embedded literally into
the title element, which is wrong.

Besides, the LaTeX footnote could never distinguish whether its HTML
alternative was a presentation or an ordinary web page.  A minor
update for htmldoc in oer-reveal-dictionaries reflects this.

Update test cases.